Fill the selected command back into the prompt when tabbing back from suggestions
For custom commands it is useful to select an earlier command, tab back into the
prompt field, and edit it there. This makes it somewhat similar to working with
your shell history.

For other types of suggestion panels we don't enable this behavior, as you can't
create arbitrary new items there that don't already exist as a suggestion.